Follow this tutorial for installing darkstat https://www.tecmint.com/darkstat-web-based-linux-network-traffic-analyzer/

+++ modules/desktop/init-scripts/11-sudoers.sh
@@ -1,7 +1,9 @@
 #!/bin/sh -eux
 
-echo "Running script sudoers.sh..."
-echo
+if [ ! -f /etc/.terraform-complete ]; then
+    echo "Terraform provisioning not yet complete, exiting"
+    exit 0
+fi
 
 # Set up password-less sudo
 echo "${DESKTOP_USERNAME} ALL=(ALL) NOPASSWD:ALL" > /etc/sudoers.d/99_user;
